chore: assistant types
