+++ src/leapfrogai_api/routers/openai/assistants.py
@@ -5,7 +5,7 @@ from openai.types.beta import Assistant, AssistantDeleted
 
 from leapfrogai_api.backend.helpers import object_or_default
 from leapfrogai_api.typedef.assistants import ListAssistantsResponse
-from leapfrogai_api.routers.openai.requests.create_modify_assistant_request import (
+from leapfrogai_api.typedef.assistants import (
     CreateAssistantRequest,
     ModifyAssistantRequest,
 )
